requirements.txt
cffi==1.9.1
pycparser==2.17
PyNaCl==1.0.1
six==1.10.0
